Change of design - we actually won't mix c and asm
The main idea behind this design decision is that assembly and C are completely different mechanisms... we want to keep things clean. C is controlled by the compiler, and assembly is controlled completely by the programmar. We do not want the mess of mixing platform specific code into intermediate code which is spat out by the compiler...